talk2postgres/requirements.txt
fastapi==0.135.1
langchain==1.2.10
langchain_groq==1.1.2
langgraph==1.0.10
python-dotenv==1.2.2
Requests==2.32.5
SQLAlchemy==2.0.36
streamlit==1.48.0
